+++ offchain/crates/solana-client-tools/src/lib.rs
@@ -3,4 +3,6 @@ pub mod instruction;
 pub mod keypair;
 pub mod payer;
 pub mod rpc;
+#[cfg(feature = "squads")]
+pub mod squads;
 pub mod transaction;